requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
av==14.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.5.0
emoji==2.14.0
faster-whisper==1.1.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
joblib==1.4.2
mpmath==1.3.0
nltk==3.9.1
numpy==2.1.3
ollama==0.4.2
onnxruntime==1.20.1
packaging==24.2
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==5.29.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
stream2sentence==0.2.9
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
